+++ packages/orq-rc/src/orq_ai_sdk/models/createchunkop.py
@@ -257,7 +257,7 @@ class CreateChunkResponseBody(BaseModel):
     updated_by_id: str
     r"""The unique identifier of the user who updated the chunk"""
 
-    id: Optional[str] = "chunk_01JQTVD4G79KFD7HPDXQ9XGSPX"
+    id: Optional[str] = "chunk_01JQV7RHAVTVDWP4NRVXP8ZREG"
     r"""Unique identifier for the element"""
 
     enabled: Optional[bool] = True