+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQTVD4G6DTPWJSFX6C303HJT"
+        "01JQV7RHAS3JX3C6TATMRSYAB5"
     )
     r"""The id of the resource"""