+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -75,7 +75,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2025-04-02T08:55:33.370Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2025-04-02T12:31:29.857Z")
     r"""The date and time the resource was created"""