+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQTVD4G5XGYCK1RD4R1BTDDJ"
+        "01JQV7RHAQT4NVF4K934V0FP1X"
     )
     r"""The id of the resource"""