+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -86,7 +86,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQTVD4G6N628V60Q9SWN54VB"
+        "01JQV7RHARKMSCW878M5WCH432"
     )
     r"""The id of the resource"""