+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -115,5 +115,5 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-04-02T08:55:31.833Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-04-02T12:31:28.307Z")
     r"""The date and time the resource was last updated"""