+++ packages/orq-rc/RELEASES.md
@@ -1089,3 +1089,13 @@ Based on:
 - [python v3.3.0-rc.27] packages/orq-rc
 ### Releases
 - [PyPI v3.3.0-rc.27] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.27 - packages/orq-rc
+
+## 2025-04-02 12:33:09
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.527.1 (2.564.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.3.0-rc.28] packages/orq-rc
+### Releases
+- [PyPI v3.3.0-rc.28] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.28 - packages/orq-rc